backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
Pillow
pytesseract
PyMuPDF
python-dotenv
python-jose[cryptography]
tesseract